requirements.txt
Click==7.0
Django==2.1.7
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.5
mccabe==0.6.1
pip-tools==3.3.2
psycopg2==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.0
pytz==2018.9
six==1.12.0
